move calendar icons
